% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odeGUTS-package.R
\docType{data}
\name{fit_odeGUTS}
\alias{fit_odeGUTS}
\title{Model calibration results for a GUTS-SD theoretical species exposed to a compound.}
\format{
A list of class \code{survFit} constructed
\describe{
\item{\code{mcmc}}{A list of mcmc chain results}
\item{\code{model_type}}{A character string containing the type of GUTS model used (here 'SD').}
}
}
\usage{
data(fit_odeGUTS)
}
\description{
Model calibration results for a GUTS-SD theoretical species exposed to a compound.
}
\keyword{dataset}
