% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odeGUTS-package.R
\docType{package}
\name{odeGUTS-package}
\alias{odeGUTS}
\alias{odeGUTS-package}
\title{odeGUTS: Solve ODE for GUTS-RED-SD and GUTS-RED-IT Using Compiled Code}
\description{
Allows performing forwards prediction for the General Unified 
          Threshold model of Survival using compiled ode code. This package 
          was created to avoid dependency with the 'morse' package that requires 
          the installation of 'JAGS'. This package is based on functions from 
          the 'morse' package v3.3.1: Virgile Baudrot, Sandrine Charles, 
          Marie Laure Delignette-Muller, Wandrille Duchemin, Benoit Goussen, 
          Nils Kehrein, Guillaume Kon-Kam-King, Christelle Lopes, Philippe Ruiz, 
          Alexander Singer and Philippe Veber (2021) <https://CRAN.R-project.org/package=morse>.
}
\references{
Virgile Baudrot, Sandrine Charles, Marie Laure Delignette-Muller, Wandrille
Duchemin, Benoit Goussen, Nils Kehrein, Guillaume, Kon-Kam-King,
Christelle Lopes, Philippe Ruiz, Alexander Singer and Philippe Veber (2021).
morse: Modelling Tools for Reproduction and Survival Data in Ecotoxicology.
R package version 3.3.1. https://CRAN.R-project.org/package=morse
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bgoussen/odeGUTS}
  \item Report bugs at \url{https://github.com/bgoussen/odeGUTS/issues}
}

}
\author{
\strong{Maintainer}: Benoit Goussen \email{benoit.goussen@ibacon.com} (\href{https://orcid.org/0000-0001-7204-7981}{ORCID})

}
\keyword{internal}
